Move filesystem to its own project.
