+++ lib/filesystem/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_compile_features(filesystem INTERFACE cxx_std_17)
+target_include_directories(filesystem
+        INTERFACE
+        ${CMAKE_CURRENT_LIST_DIR}
+)
