+++ lib/options/CMakeLists.txt
@@ -8,5 +8,4 @@ target_sources_local(options
         user_options.cpp
         option_file.cpp
         option_file.hpp
-        filesystem.hpp
 )
