+++ lib/options/global_options.cpp
@@ -1,7 +1,7 @@
 #include "global_options.hpp"
-#include "filesystem.hpp"
 #include "user_options.hpp"
 #include <constants.hpp>
+#include <filesystem.hpp>
 #include <msync_exception.hpp>
 #include <print_logger.hpp>
 #include <whereami.h>
