+++ lib/options/user_options.hpp
@@ -3,9 +3,9 @@
 
 #include <string>
 
-#include "filesystem.hpp"
 #include "option_enums.hpp"
 #include "option_file.hpp"
+#include <filesystem.hpp>
 
 struct user_options
 {
