+++ tests/test_helpers.hpp
@@ -1,7 +1,7 @@
 #ifndef _TEST_HELPERS_HPP_
 #define _TEST_HELPERS_HPP_
 
-#include "../lib/options/filesystem.hpp"
+#include <filesystem.hpp>
 
 #include <fstream>
 #include <string>
